requirements.txt
# Vision Tokenizer Playground - Requirements
# Tested with Python 3.11 and CUDA 12.1

# Core Deep Learning
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
flash-attn==2.8.3

# Transformers & Diffusion Models
transformers==4.57.1
diffusers==0.35.2
huggingface-hub==0.36.0

# Metrics
lpips==0.1.4
pytorch-msssim==1.0.0
clean-fid==0.1.35
scikit-image==0.25.2

# Utilities
einops==0.8.0
easydict==1.13
omegaconf==2.3.0
torchdiffeq==0.2.5

# Data Processing
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python-headless==4.12.0.88

# Visualization & UI
matplotlib==3.9.3
seaborn==0.13.2
gradio==6.0.0

# Configuration & Progress
pyyaml==6.0.2
tqdm==4.66.6

# Dataset
pycocotools>=2.0.6
